requirements.txt
black==24.4.2
celery==5.4.0
Django==5.0.4
django-cors-headers==4.3.1
django-deprecate-fields==0.1.2
django-oauth-toolkit==2.3.0
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-prometheus==2.3.1
django-timestampedmodel==0.1.0
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
django-rest-swagger==2.2.0
gunicorn[gevent]
phonenumbers==8.13.35
pre-commit==3.7.0
python-decouple==3.8
pytz==2024.1
redis==5.0.4
requests==2.31.0
sentry-sdk==2.0.1
